[REFACTOR]: Change prop passing method to using spread operator
