+++ app/features/Request/Body/RequestBodyMenuOption.js
@@ -1,22 +1,14 @@
 import React from "react";
 
-import KeyValueTable from "../../../components/KeyValueTable";
 import { FORM_DATA, FORM_URLENCODED, RAW } from "../../../constants/request";
+import KeyValueTable from "../../../shared/KeyValueTable";
 import useKeyValueTableStore from "../../../store/useKeyValueTableStore";
 import useRequestStore from "../../../store/useRequestStore";
 import RequestNoBody from "./RequestNoBody";
 
-function RequestBodyMenuOption() {
+const RequestBodyMenuOption = () => {
   const { bodyOption } = useRequestStore((state) => state);
-  const {
-    requestKeyValueTableData,
-    handleRequestCheckbox,
-    handleRequestKey,
-    handleRequestValue,
-    handleRequestDescription,
-    addNewTableRow,
-    deleteTableRow,
-  } = useKeyValueTableStore((state) => state);
+  const keyValueProps = useKeyValueTableStore((state) => state);
 
   switch (bodyOption) {
     case FORM_DATA:
@@ -25,20 +17,15 @@ function RequestBodyMenuOption() {
         <KeyValueTable
           type={bodyOption}
           title={bodyOption}
-          tableData={requestKeyValueTableData}
-          handleCheckboxInput={handleRequestCheckbox}
-          handleKeyInput={handleRequestKey}
-          handleValueInput={handleRequestValue}
-          handleDescriptionInput={handleRequestDescription}
-          handleAddButton={addNewTableRow}
-          handleDeleteButton={deleteTableRow}
+          {...keyValueProps}
         />
       );
     case RAW:
-      return <h1>Raw Data Input</h1>;
+      return <h1>Raw</h1>;
+
     default:
       return <RequestNoBody />;
   }
-}
+};
 
 export default RequestBodyMenuOption;
